cylkows-tanakh-epub/requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
python-hebrew-numbers==0.2.3
PyYAML==6.0.1
soupsieve==2.5
